+++ TrafficControl.DAL.RestSharp/Types/PostCaseDTO.cs
@@ -3,7 +3,7 @@ namespace TrafficControl.DAL.RestSharp
     public class PostCaseDTO 
     {
         public string ErrorDescription { get; set; }
-        public int Installation { get; set; }
+        public long Installation { get; set; }
         public ObserverSelection Observer { get; set; }
     }
 }
